+++ src/test/java/com/marginallyclever/convenience/TestStringHelper.java
@@ -1,5 +1,6 @@
 package com.marginallyclever.convenience;
 
+import com.marginallyclever.convenience.helpers.StringHelper;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -7,7 +8,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 public class TestStringHelper {
     @Test
     public void testPaddedHex() {
-        assertEquals("000001",StringHelper.paddedHex(0x1));
+        assertEquals("000001", StringHelper.paddedHex(0x1));
         assertEquals("0007ff",StringHelper.paddedHex(0x7ff));
         assertEquals("100001",StringHelper.paddedHex(0x100001));
         assertEquals("1000001",StringHelper.paddedHex(0x1000001));
